The complete game scripts.
On touch, the player object displays a dialog box to the player.
The player chooses between attacking, healing or charging their mana.
When the player makes a choice, the player object will send a message on
a private channel to prompt the villain to subtract HP. The villain's
choices work the same way.

The player loses if their health or mana reaches 0. The villain only
loses if reduced to zero hp.